pyproject.toml
[project]
name = "debate-agent"
version = "0.1.0"
description = "Multi-Agent AI Debate Arena using LangGraph + Groq"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langchain",
    "langchain-groq",
    "langchain-community",
    "langchain-core",
    "langgraph",
    "duckduckgo-search",
    "ddgs",
    "python-dotenv",
    "streamlit",
    "ipykernel",
]

[project.optional-dependencies]
dev = [
    "ipykernel",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel",
]